requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow
python-binance
python-dotenv
finnhub-python
transformers
torch
psycopg2-binary
SQLAlchemy